Code/API Deployment/requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
Cython==0.29.23
Flask==2.0.3
Flask-Cors==3.0.10
flask-ngrok==0.0.25
flatbuffers==2.0
fonttools==4.31.2
future==0.18.2
gast==0.5.3
gensim==4.1.2
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.1
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
nltk==3.7
numpy==1.21.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.1.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
python-twitter==3.5
pytz==2022.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
smart-open==5.2.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tqdm==4.63.1
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.14.0
zipp==3.7.0
